[FIX] knowledge: remove the focus on click on anchors
This commit fixes an issue where the focus of the user's mouse would be moved
to the comment box when clicking on the corresponding anchor.
This renders the edition inside of the anchors very hard to do.

Now, we removed this behavior and you only trigger the focus when clicking on
the comment box. Simplifying the edition of text inside of the anchors.

task-3790314

Part-of: odoo/enterprise#58146