[FIX] spreadsheet_edition: Fix insert button tooltip
The button "insert in spreadsheet" added in the pivot view would display
a tooltip of "0" when the button was active. The tooltip was only meant
to be present when the button was disabled due to the presence of
duplicated group bys in the pivot.

This revision fixes the tooltip presence on the button and also removes
a test that:
- had a description that did not match the actual test
- was in fact testing the same behaviour as another test -> useless
  redundancy

closes odoo/enterprise#58320

Task: 3790072
X-original-commit: 540365265cedbe03613879a16e260327f6646a8e
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>