[FIX] spreadsheet_edition: Fix insert button tooltip
The button "insert in spreadsheet" added in the pivot view would display
a tooltip of "0" when the button was active. The tooltip was only meant
to be present when the button was disabled due to the presence of
duplicated group bys in the pivot.

This revision fixes the tooltip presence on the button and also removes
a test that:
- had a description that did not match the actual test
- was in fact testing the same behaviour as another test -> useless
  redundancy

closes odoo/enterprise#58337

Task: 3790072
X-original-commit: 944bd48df133f2bc00357b5a3df3910da62914c2
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>