[PERF] stock_barcode: improve barcode view rendering with many lines
When the barcode application has many lines, it becomes slow and
unresponsive, especially on mobile devices, which are widely used with
the barcode app.

Before this commit:
Selecting a line in the barcode application triggers the rendering of
the whole view. Each line rendered calls the getter `get selectedLine`
multiple times, which is slow as it loops over all lines to find the
currently selected line.

After this commit:
Do not call the getter to check whether the current line is selected or not.
This reduces the complexity from O(n^2) to O(n).

Benchmark:
On a laptop and with 2000 lines, selecting the last line took 7.5s before,
and 600ms with this commit.

opw-3754849

closes odoo/enterprise#58560

X-original-commit: 80d9a4fb0e0c38a79aa76c9dc17bb7a35a8c83c1
Signed-off-by: Steve Van Essche <svs@odoo.com>