[FIX] planning: fix page not found errors from front-end calendar
Behaviour prior to fix:
Whenever we used any of the buttons in the planning front-end we
would get a page no found error when being redirected.

Cause:
The planning and employee tokens could not be fetched from the DOM
as we did before because the calendar view was not properly rendered
with the update to FC v6. This lead to the tokens being undefined
when clicking the buttons and thus the redirect leading to the page-not-
found errors.

Fix:
Changed the element that we mount to the FC to more specifically be the
calendar widget. That way DOM of the  calendar post render more closely
resembles what it looked like before the update to v6 and thus the buttons
work again.

task-3791067

closes odoo/enterprise#58186

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>