[MOV] appointment{_hr}: enable "conflict warning" in base appointment
When installing appointment_hr, users will get a warning in the small form
of the gantt view if there is a conflict between another appointment
and this appointment for the organizer.

We move that code to appointment so that users may get the warning
of conflicting appointments, even when there are not hr "leaves" per se.

odoo/upgrade#5800

task-3700412

Part-of: odoo/enterprise#58245