[FIX] knowledge: fix the undefined comment box in panel
This commit fixes an issue with the comments panel when creating comment.
When you have the panel opened and you want to create a comment, after logging
your message the temporary box is not removed, leading to issues like boxes not
being handled or crashes.

This is caused because when syncing the handler and the panel when creating the
comment, the insertion function was still the one from the handler meaning that
the KnowledgeComment was not using the correct function which led to those
issues.

Now when the panel is adding the comment from the handler we change the
function so that it uses the correct one.

task-3792116

closes odoo/enterprise#58433

X-original-commit: b027da13dd554181c35952a752d3ea1ff585fee7
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>