[FIX] website_knowledge: allow read access to stages of published article
Current behavior before this commit:
-> Create 1 article that has an embedded kanban and 1 article inside a
kanban(both articles are published). Now if the user opens this article
from another user - member/portal, the user gets an error like
'Something went wrong'.

Reason:
-> Access to the Kanban stage was only for the user who had access to that
article, there was no record rule defined that the stage was shown to the
member/portal user if the article was published and the article was an item
article.

Solution:
-> In previous versions there was a temporary fix that was added here(https://github.com/odoo/enterprise/pull/56230/commits/e9f63e0aa0bb51e2d686473227671ab392d56b9e)
so removing that fix and  adding the access rule which allows other users to
read stages if the article is published.

Desired behavior after this commit:
-> Kanban will appear properly with read access to the portal/member if
the article is published.

Task-3486083

closes odoo/enterprise#58272

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>