[REF] account: use _get_sql_table_expression instead of _query_get
The functions `_get_sql_table_expression` and `_query_get`
return basically the same SQL query.
The only difference is in representation.
The former returns SQL objects while the latter uses strings and lists.

The SQL object (odoo.tools.SQL) was introduced in version 17.0.
It stores the queries and (to be injected) parameters in a single object.
This helps with composing / injecting SQL queries.
It is now preferred to just compose SQL objects instead of using mogrify.

After this commit _get_sql_table_expression is used everywhere
instead of _query_get.
The latter function is removed since it is now obsolete

task-3791228

Part-of: odoo/enterprise#59267