[IMP] planning: add action buttons to the shift emails
Prior to version 15 we had action buttons in the single
shift email of the planning app. Specifically, if an open
shift was shared from the planning app, employees could get
the open shift assigned to them through a self-assign button
in the email. Similarly, if the self-unnassign setting is
enabled from the planning app's settings, employees could
unnassign themselves from their shifts directly from the
shift's sharing email through a self-unnassign button.

This PR re-introduces the two buttons to the shift emails.
The code for their logic was never removed so we only
reintroduce them in the email template.

We also change the way that the planning is shared. Prior
to this PR, we would check whether the employee was a planning
internal user by checking whether they have a coresponding
user and whether that user has group_planning_user as one of
their groups. In the planning app though there are three types
of internal users, ones who can see only the my planning menu,
ones who can see other menus and other resources, and planning
managers. The former type of planning user did not have a group.
This resulted in the planning being shared with internal users
but only opening in front-end view.

task-3791152

closes odoo/enterprise#58423

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>