[IMP] sign: allow linking sign request activity with origin record
This commit has the goal to increase the usefulness of the "request signature" activity type.
This is done by integrating the model that has the activity with the signature.request record.
Now, when a sign activity is added to a record, it is linked to any sign request that is generated from that activity.
When sending the request, a message stating that the request was sent is added in the chatter for logging purposes.
Additionally, whenever the sign.request is completed, the generated files are added in the chatter.
This makes it easier to navigate between the linked records.

task-3685648

Part-of: odoo/enterprise#55993