[FIX] planning: fix access rights errors in planning front-end
This commit fixes some access rights errors that were occuring when
clicking on buttons that would change the resource_id in the
planning front-end or from the planning automatic emails.

Behaviour prior to fix:
Clicking on "I take it", "I am unavailable" in the front-end or in
the automatic planning emails would lead to a planning rights error.

Cause:
When changing the resource_id from the front-end we do it trough sudo.
Apparently, when redirecting the user and fetching the edited slot,
the env changes, the sudo is lost, and the slot_properties field is
recomputed. The access rights error is raised from the re-computation
of that field.

Fix:
We call the field after changing the resource so it is re-computed with
sudo. Then, when re-directing, there is no-need for re-computation and
thus we don't get an access rights error.

task-3793005

closes odoo/enterprise#58280

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>