[FIX] social: improve the performances of the social controllers
This commit improves the '_get_social_stream_post' method to make it use a
complete domain instead of relying on pathing through the stream, then the
account, then the media (= 3 records fetch), to get the media type.

This will save a lot of requests as it is a central method used in the
controllers of all social media implementations.

closes odoo/enterprise#58388

X-original-commit: 9bc309a82281fcbd43bc029496bb88de7cd94782
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>