[FIX] account_reports: amount rounding tour
The amount rounding tour was broken. As it used `:contains(),
it was always checking for a substring of the current value
of a line. Since it has to wait for the change to open (rpc
call), there was a lot of indeterminism.

It is fixed by adding more robust extra triggers.

closes odoo/enterprise#58976

X-original-commit: f7f313802bb8a8c1a44404811f1528b002fe31f5
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>