[FIX] account_reports: amount rounding tour
The amount rounding tour was broken. As it used `:contains(),
it was always checking for a substring of the current value
of a line. Since it has to wait for the change to open (rpc
call), there was a lot of indeterminism.

It is fixed by adding more robust extra triggers.

closes odoo/enterprise#58285

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>