[FIX] account_invoice_extract: disable detected layout matching for CIs
The idea behind this matching is that invoices emitted by the same
supplier will often look very much alike and thus share the same layout.

Using the detected layout to identify the customer of a customer invoice
doesn't make sense since the layout only identifies the supplier of the
invoice.
This feature should only be used on vendor bills.

closes odoo/enterprise#58391

X-original-commit: b0fca800c0c0e7f59bb746840f155f45de8b425f
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>