[FIX] account_invoice_extract: disable detected layout matching for CIs
The idea behind this matching is that invoices emitted by the same
supplier will often look very much alike and thus share the same layout.

Using the detected layout to identify the customer of a customer invoice
doesn't make sense since the layout only identifies the supplier of the
invoice.
This feature should only be used on vendor bills.

closes odoo/enterprise#58312

X-original-commit: aad4ba3ad6b6a7e5dc08254ea84fccb27a351ec8
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>