[FIX] sale_subscription: display duration for subscription lines only
Versions
--------
- 16.0+

Steps
-----
1. Create a subscription SO;
2. confirm, invoice, and upsell;
3. on upsell, add a non-subscription item;
4. create invoice.

Issue
-----
On the invoice, the non-subscription item gets labeled with a duration,
as if it's supposed to be returned afterwards.

Cause
-----
When preparing invoice lines, the conditional checks whether the line
has a subscription item, or the order is an upsell. The latter part was
added in c8b79afe1bf2971a7ac7709e7165b863969d6a3d to store subscription
id, start_date and end_date on upsell invoice lines.

Solution
--------
Add an additional conditional to format the description, only displaying
duration for items that are subscription-based.

opw-3687626

closes odoo/enterprise#59897

X-original-commit: 1317648e9296c235db82b607e380f13a2e99c1ae
Signed-off-by: Levi Siuzdak <sile@odoo.com>