[IMP] sale_subscription: improve close reasons UI change
In this commit, the user interface for the close reasons form is enhanced.
the field string visible_in_portal is modified to Selectable in Portal and
additional guidance is provided for this field. A new condition is introduced,
ensuring that if the field is set to false, the subsequent section becomes
invisible otherwise, it remains visible.

task-3776667

closes odoo/enterprise#58401

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>