[FIX] web_studio: reportEditor meaningful display when loading error
Since commit [1], the report editor never crashes at loading time.
This is intended to ease the migration process, or report recovery in general, where the operator could
reset the views, explore the diff of some views, or edit their plain XML.

However, commit [1] did not offer a way to actually see the error in the reportEditor itself. This was on purpose
as the lambda user would not know what to do anyway with that information.

This commit aims at offering that piece of information in case we are in debug mode.
The display of the error allows to open the incriminating views in a dialog, and see some
basic information on them, as well as, obviously, the error message and type.

opw-3735973

[1]: 62c167f87176674cbabad976d93eeb717f73df7a

closes odoo/enterprise#58491

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>