[IMP] *: gantt: variable time range
The main goal of this task is to allow the user to work on a chosen time
range via two date pickers. This will ease the edition of records by making
possible to move, resize, create, edit records on time periods that are not
arbitrarily fixed to a given day, week, month, or year. Two new scales
are also available: week_2 and month_3. The principal effect of scales is
now to determine the duration of the gantt columns ("hour", "day", "month")
and their widths. The scale can be chosen via a slider.

We also have added some new buttons:
- two buttons for sorting the rows by the time their first pill occur
  (i.e. groups can be sorted by date_start:min asc or desc)
- a button to change the display mode ("dense"/"sparse") introduced in
  https://github.com/odoo/enterprise/pull/58338

Note that for performance reason, we do not allow time range bigger than
10 years. We also forbid to load more than 1000 records. In case that
limit is reached a notification is displayed. This has been done in order
to keep the number of rows/columns in the css grid within the bounds
allowed by Firefox (https://bugzilla.mozilla.org/show_bug.cgi?id=1336679).
Note that we do not render all loaded records. The gantt view is now also
virtualized in the horizontal direction and we only add to the css grid
the rows/columns  effectively needed to display the visible elements.

In mobile mode, the date selection is done in a dialog instead of a
popover. We plan to extend that behavior everywhere.

Task ID: 3601456

closes odoo/enterprise#55618

Related: odoo/odoo#162017
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Bruno Boi <boi@odoo.com>
Co-authored-by: Julien Carion <juca@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>
Co-authored-by: Romeo Fragomeli <rfr@odoo.com>
Co-authored-by: Pierre Pulinckx <pipu@odoo.com>
Co-authored-by: Luca Vitali <luvi@odoo.com>