[FIX] account_reports: properly handle aggregation bounds referenced by cross_report expressions
To reproduce:

- Create a report with just one line, with code line_1_1. Give it an aggregation expression, cross_report, with formula "line_2_1.balance"

- Create another report, containing a single line with code line_2_1. Give it two expressions:
    - One aggregation, with formula="14.0" (hardcoded fixed value), and subformula "if_other_expr_above(line_2_1.dudu, EUR(0))"

    - Another expression of any engine labelled 'dudu'

==> Try opening the first report; it loads forever.

This is because the bound expression is only searched in the expressions of the active report ; not in the expressions loaded from other reports. In our case, the aggregation needing the bound is from the second report, so the bound belongs to its report as well. The looping occurs because that value, since it's not present in the first report's evaluated data, is considered as not yet computed, and the computation is continuously delayed at the end of the treatment queue, causing it to never become empty.

closes odoo/enterprise#59236

X-original-commit: aba82d9c4fb67adabf3ddf3c883e9022816aeb8e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>