[FIX] pos_preparation_display, pos_self_order_pdis: conf page behavior
The confirmation page wasn't displayed and the order wasn't sent to the
preparation display when confirming an order with a total of 0.

Now the confirmation page is displayed and the order is send to the
preparation display when the order is confirmed regardless of the total
amount.

closes odoo/enterprise#58348

Related: odoo/odoo#157028
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>