[IMP] industry_fsm: display empty stages if there is only one FSM project
Before this commit, the pipeline stages displayed in Field service in the
kanban/list and the other views were only stages containing at least one task.

After this commit, all stages are shown (even empty ones) if there is only
one 'field service' type of project created. If more, we fallback on the
current behavior (i.e. only showing stages containing tasks).
Moreover, we now hide the project_id field in the views if, again, there is
only one FSM project.

The rationale behind this commit is that we want to mimic the behavior we
have when displaying the tasks of a specific project from the Project module.

task-3781207

closes odoo/enterprise#58375

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>