[IMP] room: add end current booking button
Purpose:
--------
Add a button in the front end view that allows users to end the current
booking (to make the room appear as free), as the only way to do it for
now is to delete the booking, which is not suitable if one would like to
track the bookings in the backend view.

This commit also rewords the "book now" button to "start now" and the
"schedule booking" button to "Add a booking" because it was observed
that users were sometimes confused with these two buttons.
The "booked room" icon has been changed to not use the same as in the
new "end booking "button. Colors of the buttons have been updated.

Task-3792497

X-original-commit: 6117ed1825047b3a6d7a99adcf8d80d2eb8cffe5
Part-of: odoo/enterprise#61485