[IMP] account_reports : Adjust the text wrap on popover
The text words were wrap when end of line, they shouldn't. The popover appear in front of the rapport and hide the content, I moved the popover on the right and change the way the words are wrapped. Same for the carryover popover.

Using a fixed value for the popover width because using % don't work as expected.

task id : 3792849

closes odoo/enterprise#58406

Signed-off-by: Quentin De Paoli <qdp@odoo.com>