[IMP] portal,* : decrease editor option label width
This commit reduces the length of the options labels when accessing the
web editor within portal.

Prior to this commit, the labels were using a weird wording and were
truncated, making it difficult to identify each option.

To fix this issue, we remove the "show" word from all the labels.

part of task-task-3703330

Part-of: odoo/enterprise#56830