[IMP] *: remove cids from URL
*: [account_reports, documents, social, stock_barcode, web_studio]

The aim of this commit is to remove the companies IDs from the URL, and
use only the cookies to store the selected companies.
This commit will also improve the handling of access errors. If a user
encounters an access error, if one of its allowed companies has access
to the record, the company is added to the selected companies and the
record is reloaded.

The purpose of this is double: to have more human-readable URLs (part of
the task id : 3557575), and to avoid modifying the current selected
companies when clicking on a shared URL.

Note that this commit also modify the way the companies IDs are encoded
on the cookies. Before this commit, the IDs were separated by a comma
(",") and they were encoded as \054 by the server side but not by the
client side. Now, we use a dash ("-") to have the same encoding in both,
server side and client side.

task-id 3792021

closes odoo/enterprise#58513

Related: odoo/odoo#157399
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Damien Bouvy <dbo@odoo.com>