[FIX] l10n_mx_edi: external test fails every time
The test for `sw` fails every time because we don't have identifiers.
We don't have any, so we just get rid of the test.
Also added the message to the error, so we can know why it failed.
(Other tests failed some time, but we can't reproduce and know why it
failed).

Linked to runbot error 25908

closes odoo/enterprise#58497

X-original-commit: 43d84da079ebcfc193bd1ebacebecc8fb65e6741
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>