[FIX] whatsapp_event: Include message to unregistered attendee in total sent count
Before this commit the Whatsapp Sent Count (located in the Communication
tab of the event view) wasn't including messages sent to the attendees who
weren't confirmed.

[This commit changes]
Event mail scheduler will include seats_uncofirmed into calculation of
the mail_count_done which is the count of emails sent. This is
reflecting the actual number of messages sent.

[Reproduce]
- Install whatsapp_event
- Create event E starting in time T (Events/Events)
- Add an "Unconfirmed" attendee
- Add an E-Mail in the Communication tab of the event E
	- Set Interval time > T,
	- Template: "Email Reminder"
	- Trigger "Before the event"
- Run scheduled Action sending whatsapp (Scheduled Actions / Event: Mail Scheduler)
- BUG: message sent, but not included in the sent count in the communication tab of the event E

Related to the commit in the odoo community called:
[FIX] event: Include unregistered attendee emails in total sent count

opw-3693626

closes odoo/enterprise#60065

X-original-commit: c0d36d6b6b749bf431be6ff84ac1cc91f540c650
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>