[FIX] hr_payroll_attendance: get linked attendances
Steps to reproduce:
-------------------
- being in Europe/Brussel timezone;
- create a contract based on attendance from 1st to 31st January;
- create two attendancescfot the employee linked to this contract:
    - 5th January from 8am to 4pm for example (A)
    - 31st January from 00:30am to 8:30pm (B)
- create a payslip for the employee from 1st to 30th January;

Issue:
------
The attendance count on the smart button on the payslip and
the number of records on the list view (when the smart button is clicked)
are not the same.

Cause:
------
The domain that determines attendance uses a condition on the `check_in` field,
which is a datetime field, and compares it with a date.
For attendance B, we will compare 2024-30-01 23:30:00 with
2024-30-01 23:59:59.

As a result, attendance B will validate this condition.

Solution:
---------
Make this information consistent by using the same logic
and taking account of timezones.

Add a method that allows you to retrieve all attendances linked to a payslip.
Use this to generate the domain that will select the correct records
(based on their id) in the list view.

Note:
-----
`_read_group` manages timezones and therefore
`check_in:day` for attendance B returns 2024-31-01.

opw-3794161

closes odoo/enterprise#58587

X-original-commit: 209914d787d657f103da845a91175c968823818e
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>