[FIX] sale_subscription: protect discount field during confirmation
Versions
--------
- saas-16.4+

Steps
-----
1. Create a pricelist wich shows discounts;
2. create a SO using this pricelist;
3. add a non-subscription-based product;
4. modify its discount;
5. confirm the SO.

Issue
-----
The manually modified discount gets overwritten by the pricelist one.

Cause
-----
Commit 71f15a9af4063bfb960b4706c3f237e69dd28e16 started protecting the
discount field in `action_confirm` to protect it from recomputation when
a start date would be added. This was adequate for earlier versions,
but commit 4ac62c700826ce1b38307641f91b7c9ce4f91d54 added a line setting
`subscription_state` to `False` for any SO that isn't an upsell or
subscription. As this field is also a dependent of
`_compute_subscription`, a recompute is triggered.

Solution
--------
Move the `with self.env.protecting` clause up, before the for-loop which
modifies `subscription_state`.

opw-3646915

closes odoo/enterprise#58704

X-original-commit: 8225a3c811ef0ec70904711a39687bf29c87c629
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>