[FIX] l10n_pe_edi: fix Detraccion amount
Steps to reproduce:
- Install Accounting and l10n_pe_edi
- Switch to a Peruvian company (e.g. PE Company)
- In Contacts, Configure a bank account (Banco de la nación - BANCPEPL)
  for PE Company
- In Accounting settings, run the automatic currency rates service ([PE] SUNAT)
- Create a product: (e.g. Product X)
  * Sales Price: 990.00
  * Withhold code: Arrendamiento de bienes muebles
  * Withhold Percentage: 10.00
- Create an invoice:
  * Customer: [a Peruvian contact] (e.g. Comercial Constructora los Patitos S.A.)
  * Operation Type: [1001] Operation Subject to Detraction
  * Payment terms: End of Following Month
  * Journal in: USD
  * Invoice Lines:
    - Product: Product X
    - Account: 7012100 Merchandise - Merchandise - Local sale - Third parties
    - Price: 990.00
    - Taxes: 18%
    - EDI Affect. Reason: Taxed- Onerous Operation
- Confirm the invoice
- Process to EDI service

Issue:
The remaining amount after deducting the withholding is not correct in the
generated EDI document.
The total is 990.00 + Taxes (18%) = 1168.20
The withholding is 1168.20 * 10% = 116.82
The remaining amount should be 1168.20 - 116.82 = 1051,38
However, the remaining amount set in the EDI document is 1052,20, which is
not correct.

Cause:
The rounding used to compute the withholding amounts are not correct.
The amount in the selected currency should contain 2 decimal digits (precision_rounding
should be 0.01, instead of 2).
Also, the Detraction amount should be declared in PEN currency in the EDI document
and should not contain the decimal part (precision_rounding should be 1, instead
of 2).

opw-3747620

closes odoo/enterprise#59099

X-original-commit: 003e891f44d6f32fd2a9ddfb60b464ad83a3d4ad
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>