[IMP] account_reports: solve 0-division issues in aggregations
In an aggregation, when a 0-division happens in a formula, the result of the formula will be set to 0. This way of handling 0-division is counter-intuitive and could potentially create errors in the reports.

This commit takes the default 0-division safety off, which forces the dev to think about- and fix these issues. To enable the safety, a developer can enter 'ignore_zero_division' as a subformula on an aggregation.

The 'ignore_zero_division' subformula was added to existing reports and testcases to stay consistent with earlier versions.

closes odoo/enterprise#58472

Task: 3774403
Related: odoo/odoo#158791
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>