[FIX] pos_preparation_display: split tests
The `pos_preparation_display` launches tours both in the `point_of_sale`
and in `pos_preparation_display` interfaces.

Previously, it was importing all the tours files from pos, which is
not needed and which lead to erros. For example if one of the tours in
pos has a dependency in a different folder, than the preparation display
would have to bring this dependency too.

The solution is to split the test files in the `pos_preparation_display`
module into `point_of_sale` and `pos_preparation_display` folders.

closes odoo/enterprise#58445

Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>