[IMP] l10n_cl_edi_pos: add bidimensional code on receipt
This commit adds the bidimensional code on the receipt, it also adds to
the receipt header the requirements for considering the receipt as a
valid electronic document.

It also changes the code shown on the credit note when refunding a
previous invoice. Before the commit, it was showing
the full name (eg. BEL 00001), but we actually need to have only
the number (eg. 00001) as it is done when refunding from the backend.

closes odoo/enterprise#58479

Task-id: 3747828
Related: odoo/odoo#159671
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>