[FIX] l10n_pe_edi_stock: Fix failing tests
1. `test_send_delivery_guide` was failing because the picking vehicule
was not set up properly in the test. After this was fix, another issue
appeared: the company needed proper credential to connect to the SUNAT.
Since those were not available, the function connecting to SUNAT has
been patched to simulate a basic answer from SUNAT.

2. `test_generate_delivery_guide`'s expected document has been updated
to include the vehicule set up in the previous step.

3. Since there is no more call to an external API. The `external` and
`l10n_external` tags are not necessary anymore. This allows to group
both tests under the same class.

task-3718645

closes odoo/enterprise#63291

X-original-commit: 6e79ee004b847b7ff90143fa5fdc45f585d30cad
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>