[FIX] l10n_dk_reports: csv export header
This commit changes the danish csv report's header to match with the
correct Danish documentation requirements.

The Danis government's documentation for the annual report accounting
file seems to have disprepancies between the PDF and the website.

We got a feedback that the generated CSV did not follow the requirements
given for regnskab basis at the following link:

https://erhvervsstyrelsen.dk/vejledning-upload-af-regnskabsfil-i-regnskab-basis

Based on the documentation chapter 3 & 4, the top row of the csv file
(header) should consists of:

KONTONUMMER_20230131, KONTONAVN_20230131, and VAERDI_20230131

closes odoo/enterprise#58706

Task-id: None
X-original-commit: 9c0c2e01a84af18fd3d11f46ec506be4d0fa26a2
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>