[FIX] sale_subscription: reset canceled renewal quote to draft
Before this commit, when a salesperson canceled a renewal quotation, and
reset it to draft, the subscription_state '2_renewal' was lost.
It would affect the sale_order_log upon confirmation of that renewal and
the sale.order.log would not be correct. In that case, there would be a
missign transfer log.

Taskid: 3651477
Part-of: odoo/enterprise#58490