[FIX] industry_fsm_sale: access products in multi company environment
Steps
-----
1. Have field services and sales installed
2. Create a new company
3. Create a new product belonging to the new company
4. Log in under new company
5. Field service > click the "Product" smart button in a task with no sales order
** Read access error when accessing a page with a product from new company **

Issue
-----
In a multi-company environment, accessing the catalog of products in field service is only possible if there are no products belonging to a different company than the user's main company.
Note: this issue only arises if there's no sales order linked to the task. If an order already exists (for example by creating a sales order item in the task form), its associated company will be used.

Fix
---
We use the tasks's company to have the correct allowed_company_ids in the context.

opw-3752311

closes odoo/enterprise#60279

X-original-commit: daf3226b531ebb28ded0b8fb2055e6913279b152
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>