[FIX] approvals: Use purchase description in approval line
In Approvals, you can create Approval Request where you select
Products (in demo data : Create RFQ).

When you do that, on the request creation itself, you'll be able to pick
products and add it in lines. There's a column description, but that column
is not populated correctly, it's just a copy of the name.

It should be the description from the product, purchase tab.

closes odoo/enterprise#59133

Taskid: 3794627
X-original-commit: 52c5831ffa78b98209a809d285945b0526c5b3e5
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>