[FIX] account: allow transaction_details to be decoded as a JSON.
When using the 'Find missing transactions' from account_online_synchronization
module, the `transaction_details` field gets encapsulated in `<p></p>` HTML tags

This is caused by that field being declared as a HTML field in the transient
model `account.bank.statement.line.transient`, but then as a JSON field in the
`account.bank.statement.line` model.

That encapsulation causes Python JSON decoder to fail to decode from
`transaction_details` when called in `_format_transaction_details()` in the bank
reconciliation widget, resulting in a traceback

Removing the HTML tags after the call to `read()` in the transient model allows
the field to be correctly decoded.

opw-3783078
opw-3773454
opw-3772561
and more.

closes odoo/enterprise#59708

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>