[FIX] sale_subscription: don't loose context
Since fixed introduced in https://github.com/odoo/enterprise/pull/57590,
the context is lost when we call 'update_existing_subscriptions'.

As we put a dictionnary returned by '_upsell_context' in the
'with_context' to call write on the subscription, it replace the current
context by the new one.

As we just want to add values to the context and not replace it, we get
the key/values as named arguments to update the current context and not
replace it

closes odoo/enterprise#58553

X-original-commit: 1b51aed8189402a8fc0179e883116735f758f024
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>