[FIX] whatsapp: correctly copy variables when copying templates with buttons
Fix the whatsapp template buttons and/or variables which are not duplicated
correctly at template duplication.

At template duplication the template variables from the header body and buttons
are duplicated but some variables are duplicated twice or wrongly. To prevent
duplicating twice the buttons variables, we now copy only variables linked
to the template when copying the template, and let the variables linked to
buttons be duplicated with the button copy.

Task-3680030 (WhatsApp: Fix buttons / template duplication)

X-original-commit: odoo/enterprise@a4e8c60a6314dd3528bf39b7031905740235b311
Part-of: odoo/enterprise#59987
Co-authored-by: Amélie Dieudonné <amdi@odoo.com>