[FIX] whatsapp: correctly copy variables when copying templates with buttons
Fix the whatsapp template buttons and/or variables which are not duplicated
correctly at template duplication.

At template duplication the template variables from the header body and buttons
are duplicated but some variables are duplicated twice or wrongly. To prevent
duplicating twice the buttons variables, we now copy only variables linked
to the template when copying the template, and let the variables linked to
buttons be duplicated with the button copy.

Task-3680030 (WhatsApp: Fix buttons / template duplication)

X-original-commit: 214a22af27c173a71e4509c8d404224ae63afb03
Part-of: odoo/enterprise#60098
Co-authored-by: Amélie Dieudonné <amdi@odoo.com>