[FIX] l10n_be_reports: Fix wrong VAT number in VAT Listing
The aim of this commit is handling correctly the
VAT number in the Partner VAT Listing. Before, to remove
the BE in the vat, we removed the 2 first characters of
the vat, it causes an issue when the VAT number doesn't
start with BE. This fix is using the _vat_to_bce function
(from the same module) which convert (if necessary) the VAT
into the BCE number.

no task id

closes odoo/enterprise#58631

X-original-commit: fcb81ce1a9dbb3ee8a7be673db9fb068efb272ee
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>