[FIX] web_studio: ButtonBox: do not display More for one button
This commit is the counterpart of odoo/odoo#156241 where we rework
the logic of computing the visible buttons and those put in a
dropdown, in the button box. There's an override in studio to
bypass the invisible filter when we have toggled the "Show Invisible"
option, so this override needed to be adapted. The fun part is that
this was actually not working, because the viewEditorModel wasn't
made reactive, so toggling the option didn't produce a re-rendering.
Now it does.

Task 3778382

closes odoo/enterprise#58607

X-original-commit: 18d02f4a9ddfb4bfd8235ba2147445ff0a2cb6c8
Related: odoo/odoo#157594
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>