[FIX] l10n_be_soda: fix the soda file detection
Bug
===
When we send a SODA file by email, it is not detected as a SODA file.

Technical
=========
Since odoo/odoo@82142475f70517045f1fbbd700e202b8dd0a522b the XML files
sent by email are imported as plain text. But the check for the SODA
file check only the mimetype XML.

Task-3792364

closes odoo/enterprise#58781

X-original-commit: 84c7126e69b2c96119a8a42507cecca821f3f248
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>