[IMP] website_sale_reting: display disclaimer in tooltip
Prior to this commit, there was a disclaimer as a title attribute in the
rental filter. But this message is too important to be displayed as a
regular title attribute.

This commit highlights this message by displaying it in a tooltip
available when hovering an "info" icon when the period is set.
For the mobile menu (the offcanvas filter menu), it displays the message
in an alert.

task-3637980

Part-of: odoo/enterprise#55520