[IMP] room: avoid flicker on quick booking
Purpose:
--------
When using the quick booking feature in the frontend view, there is
currently a flicker when the main view is restored after booking the
room.
This flicker occurs because the view is restored before receiving the
booking creation notification through the bus service.

With this commit, instead of relying on the bus notification to add the
bookings created from the current device, they are added after that the
creation/update RPCs are resolved, which allows to restore the view
after that the bookings have been added.

During these RPCs, the ui is now blocked so that the user has a feedback
of its click on the buttons.

Task-3704495

Part-of: odoo/enterprise#58572