[IMP] room: show booking when re selecting it from sidebar
Purpose:
--------
In the frontend view, when editing a booking and re-clicking on it in
the sidebar, currently nothing happens.

With this commit, it will discard the current modifications and recenter
the view on the booking being edited, by forcing the bookingToEdit prop
to be a new object when clicking on the booking in the sidebar, which
causes the component to rerender itself.

Task-3704495

Part-of: odoo/enterprise#58572