[FIX] project_enterprise: change default behavior of Plan dialog
In the Plan dialog of the project tasks gantt view, you can choose to
either plan your tasks according to the selected timespan with the
"Select" button, or you can choose to Auto Plan them with the "Auto
plan" button. However, when you click on a single line, it gets auto
planned, which shouldn't happen, it should be planned with the logic of
the "Select" button. Plus there was a traceback when trying to plan only
tasks with no assignees, because we would do a division by zero without
checking beforehand that it mustn't happen. This PR aims to fix these
problems.

closes odoo/enterprise#58571

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>