[FIX] account_intrastat: Handle correctly UK case
The aim of this commit is handling properly the UK
case when we unfold lines in the intrastat report.
For some reasons, instead of using GB , we have to
use XU for the product origin country code.
The problem was that we used this country code to find
children lines when the user unfold the Dispatch/Arrival
line. As the country code XU doesn't exist, for all these
lines, we don't get any children lines.
Now, before we build the ID (that we'll use after to fetch
children lines), we change the value for the right country
code.

no task id

closes odoo/enterprise#58682

X-original-commit: 2015bf284bdc8c3fc54c30b1298d2b94699428fc
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>