[IMP] planning(_contract, _holidays): extending and refactoring recurrency tests
This commit extends the tests performed for the recurrent shift
feature of the planning app. This features was recently improved to
take into account contracts, holidays and work hours when generating
shifts. The tests introduced in this commit are ensuring that these
new functionalities are performing as they should.

Also in this commit, we are performing a small refactoring of the
tests in the planning, planning_contract, planning_holidays modules.
Some things were redundant and there was a lot of code duplication
which now has been reduced.

task-3780480

closes odoo/enterprise#59035

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>