[FIX] l10n_be_hr_payroll: avoid using variable in lambda
Since the code is called in a safe_eval, instead of in the codebase in https://github.com/odoo/enterprise/pull/32970
We cannot use a variable in a lambda as it raises a NameError on the
employee variable called in the filtered function.

So to avoid the filtered function we loop on the leave and aggregate
manually by employees.

X-original-commit: 6965e521fffca47b5e0f178f5215cc85a251ee76
Part-of: odoo/enterprise#58671