[FIX] account_asset: wrong board asset with one period
When having an asset beginning at the beginning of a period
and having one period, the board values are wrongly computed
although the depreciation move lines are good.

Steps:

- Create an asset with acquisition date set at the first day of a month
- Set the period to 1 month
- Set the same account as fixed asset account and depreciation account
- Compute board and show depreciation moves tab
-> Depreciation value, remaining value and cumulated depreciation are
   wrong

The reason is that we enter in a condition meant for set the value of
closing entry when needed, which is not the case here.
As there is no way, at the moment we enter this condition, to check if
the asset is ment to be closed (state is still set to 'open'), the
solution is to check the number of line, which is 3 in the case of
a disposal move or 4 for a sale.

opw-3750254

closes odoo/enterprise#59265

X-original-commit: 606a14656fa3d6fba4850e8b886b09c5832dc6af
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>